requirements.txt
discord.py>=1.7.3
discord-pretty-help>=1.3.3
DiscordUtils>=1.3.4
rich>=12.0.0
coloredlogs>=15.0.1